84764 Moving send_notification to travel claim base worker
